requirements.txt
python-dotenv
pydantic
contextlib
psycopg2
pymongo
"fastapi[standard]"
bson